Fix for crashing E-RABSetupRequest with E-RAB ID > 12
Since number of Logical Channels are 0 to 10 (both inclusive) and only 3-10 are used for DRBs.
And, maximum DRBs an UE can have is 11 at any time. Therefore there is need to handle these DRBs
per UE with Logical Channels from 3 to 10. This commit address this issue by not relying on simple
calculation of LCID based on E-RAB ID (i.e. LCID = E-RAB - 2) or DRB ID. But rather keep track
of used LCID and assign unused ones when a new bearer needs to be created. If all LCIDs are used up
then eNB will respond to CN with a cause mentioning radio_res_not_available